[backend] Created package contract and relocated classes for improved organization and modularity
