+++ openex-api/src/main/java/io/openex/scheduler/jobs/ComchecksExecutionJob.java
@@ -5,6 +5,7 @@ import com.fasterxml.jackson.databind.ObjectMapper;
 import com.fasterxml.jackson.databind.node.ObjectNode;
 import io.openex.config.OpenExConfig;
 import io.openex.contract.Contract;
+import io.openex.contract.ContractService;
 import io.openex.database.model.*;
 import io.openex.database.repository.ComcheckRepository;
 import io.openex.database.repository.ComcheckStatusRepository;
@@ -12,8 +13,10 @@ import io.openex.execution.ExecutableInject;
 import io.openex.execution.ExecutionContext;
 import io.openex.injects.email.EmailContract;
 import io.openex.injects.email.EmailExecutor;
-import io.openex.service.ContractService;
 import io.openex.service.ExecutionContextService;
+import jakarta.annotation.Resource;
+import jakarta.transaction.Transactional;
+import jakarta.validation.constraints.NotNull;
 import org.quartz.DisallowConcurrentExecution;
 import org.quartz.Job;
 import org.quartz.JobExecutionContext;
@@ -22,9 +25,6 @@ import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.context.ApplicationContext;
 import org.springframework.stereotype.Component;
 
-import jakarta.annotation.Resource;
-import jakarta.transaction.Transactional;
-import jakarta.validation.constraints.NotNull;
 import java.time.Instant;
 import java.util.List;
 import java.util.Map;
